`sampled_median_filter` now works for multiple 2d images within a larger structure
Added an argument to specify dimensions for filtering, and made required changes within the function to support this.
Included tests using stacked images with both numpy and xarray.

Motivation for this was to be able to filter a DataArray containing multiple filters, dithers and detectors.
This feature will be used by banding correction.